refactor: AdminInviteService를 Feign Client 방식으로 변경(#35)

+++ guideon-admin-bff/src/main/java/com/guideon/guideonbackend/domain/zone/dto/ZoneResponse.java
@@ -25,7 +25,7 @@ public class ZoneResponse {
     @JsonProperty("zone_type")
     private String zoneType;
 
-    private Integer level;
+    private Short level;
 
     @JsonProperty("parent_zone_id")
     private Long parentZoneId;
